requirements.txt
aiogram>=3.27,<4
aiosqlite>=0.20,<1
aiohttp>=3.9,<4
python-dotenv>=1.0,<2
